fix: don't report scroll direction change when velocity is zero
Currently when the animation curve reaches a point where its velocity
is zero, the AnimatedScrollActivity will report that the user scroll
direction has changed to idle, which is not necessarily true, because
Bezier curves for example can have multiple zero velocity points over
the course of the animation curve. Moreover, the ScrollNotification
docs state that the notifications should have the following lifecycle:

 * A [ScrollStartNotification], which indicates that the widget has
   started scrolling.
 * Zero or more [ScrollUpdateNotification]s, which indicate that the
   widget has changed its scroll position, mixed with zero or more
   [OverscrollNotification]s, which indicate that the widget has not
   changed its scroll position because the change would have caused its
   scroll position to go outside its scroll bounds.
 * Interspersed with the [ScrollUpdateNotification]s and
   [OverscrollNotification]s are zero or more [UserScrollNotification]s,
   which indicate that the user has changed the direction in which they
   are scrolling.
 * A [ScrollEndNotification], which indicates that the widget has
   stopped scrolling.
 * A [UserScrollNotification], with a [UserScrollNotification.direction]
   of [ScrollDirection.idle].

This commit brings the notification lifecycle of AnimatedScrollActivity
in line with those expectations.