feat: enumerate types of scroll event sources
Chromium's default scroll animation algorithm uses different methods
of calculating animation durations based on the origin of scroll events.
For programmatic scrolls, the duration is roughly proportional to the
scroll offset, while for mouse wheel scrolls, the duration is also
roughly proportional, but to the inverse of the scroll offset. For
keyboard scrolls the duration is constant.

Passing the type of scroll events to the animation factory allows
customizing animation behavior accordingly.

We currently only support pointer-based scroll events, but this feature
is a stepping-stone on the way to adding support for keyboard and
programmatic scrolling.