fix: stop processing a tick immediately if the activity was disposed
Up to this moment, we only prevented our activity from going idle
when we detected that it was disposed. To prevent any other unwanted
side effects such as calls to setPixels or updating the user scroll
direction from an already disposed activity, let's stop processing
the tick immediately if we know the activity was disposed.

This is speculative though, when the dispose method disposes of the
ticker, then at least in theory no more ticks should arrive, so we
should probably investigate if this flag is really needed.