docs: add a todo comment for migrating away from datetimes
Relying on DateTime.now() to then measure time deltas is not a great
idea, as the system clock might not be stable. For example there might
be a Network Time Protocol client running on the system and might be
making small adjustments to the system clock, which might result in
time seemingly going faster, slower or even going backwards.