feat: allow updating the target of a programmatic scroll animation
Change the behavior of AnimatedScrollPosition.animateTo, so that for
animations with a ScrollAnimatorCurve, the target of a currently
running scroll animation is updated instead of starting a new animation.
This makes the transition between two overlapping programmatic scrolls
smoother, by making the animation velocity continous, which was suddenly
dropped to zero when a new programmatic scroll was issued while another
one hasn't finished yet before.