feat: implement scrolling using arrows or pg-up/down keys
Implement an AnimatedScrollAction as a subclass of Flutter's
ScrollAction which differs from this superclass only by the way it
performs the animation to a target value. When the Scrollable found
for the context provided to the action is managed by an
AnimatedScrollPosition, the keyboardScroll method is called instead of
ScrollPosition.moveTo, which results in an AnimatedScrollActivity
being initiated instead of a DrivenScrollActivity.